Fix new folder dialog populated with previous folder name
