+++ lib/osf-components/addon/components/files/menu/create-folder/template.hbs
@@ -8,7 +8,7 @@
     </dialog.heading>
     <dialog.main local-class='dialog-main'>
         <FormControls
-            @changeset={{this.changeset}}
+            @changeset={{@changeset}}
             as |form|
         >
             <form.text
@@ -26,8 +26,8 @@
         </OsfButton>
         <OsfButton
             @type='primary'
-            @onClick={{perform this.createFolder (hash onSuccess=dialog.close)}}
-            @disabled={{this.shouldDisableButtons}}
+            @onClick={{action @createFolder (hash onSuccess=dialog.close)}}
+            @disabled={{@shouldDisableButtons}}
         >
             {{t 'osf-components.files-widget.create'}}
         </OsfButton>
