--- lib/osf-components/app/components/files/menu/create-folder/component.js
@@ -1,1 +1,0 @@
-export { default } from 'osf-components/components/files/menu/create-folder/component';
