+++ lib/osf-components/app/components/files/menu/create-folder/template.js
@@ -1,0 +1,1 @@
+export { default } from 'osf-components/components/files/menu/create-folder/template';
